fixed "tests not received" errors
